[2단계 - DB 복제와 캐시] 콜리(김건우) 미션 제출합니다 🥦  (#108)
* feat: 회원 쿠폰 발급 기간 검증 추가

* feat: redis 의존성 추가

* feat: redis 설정

* feat: cache 설정

* refactor: 간접참조로 변경

* feat: memberCoupon 발생 기능 구현

* feat: memberCoupon service 캐싱 구현

* chore: 패키지 이동

service > service.db

* feat: service 애너테이션 추가

* refactor: key hash값을 애너테이션으로 대체

* refactor: key 값을 직접 명시

* chore: ddl 설정을 create-drop으로 변경

* docs: init.sql 수정

* refactor: 불필요한 redishash 제거

* feat: objectmapper에 datetime 모듈 추가

* feat: 직렬화를 위한 getter 추가

* feat: cache test 추가

* chore: CouponController 삭제

* refactor: dto를 반환하도록 변경

* feat: member domain 구현

* feat: getter 추가로 jsonDetect 제거

* refactor: start_at column 애너테이션 추가

* refactor: member 기본 생성자 접근 제어자 변경

* fix: member_coupon 검증 로직 에러 개선

* chore: 캐시 이름 상수화

* docs: init.sql문 문법 수정

* docs: array를 포함하도록 수정

* refator: memberService 조회 로직 개선

* test: test teardown 메서드 수정

* refactor: 저장되는 키 값 변경

coupon > result

* refactor: 발행량 검증 순서 변경

* feat: 새로운 쿠폰 발행 시 캐시 삭제

* chore: field로 직렬화를 인식하도록 수정

* feat: coupon 값을 getter로 가져오도록 변경

* feat: couponTest 이전에 캐시 비우는 작업 추가

* feat: redisConfig 설정 변경

* test: issue시 evict 로직 테스트 추가

* chore: redis config 허용 범위 변경

---------

+++ src/main/java/coupon/domain/coupon/DiscountRatio.java
@@ -21,7 +21,7 @@ public class DiscountRatio {
         }
     }
 
-    public double getRatio() {
+    public int getRatio() {
         return ratio;
     }
 }
